+++ edc-c-client/protobuf/src/google/protobuf/unknown_field_set.h
@@ -115,7 +115,7 @@ class LIBPROTOBUF_EXPORT UnknownFieldSet {
   bool ParseFromZeroCopyStream(io::ZeroCopyInputStream* input);
   bool ParseFromArray(const void* data, int size);
   inline bool ParseFromString(const string& data) {
-    return ParseFromArray(data.data(), data.size());
+    return ParseFromArray(data.data(), (int)(data.size()));
   }
 
  private:
@@ -197,7 +197,7 @@ inline void UnknownFieldSet::Swap(UnknownFieldSet* x) {
 }
 
 inline int UnknownFieldSet::field_count() const {
-  return (fields_ == NULL) ? 0 : fields_->size();
+  return (fields_ == NULL) ? 0 : (int)(fields_->size());
 }
 inline const UnknownField& UnknownFieldSet::field(int index) const {
   return (*fields_)[index];
